refactor(account-settings): separate getAccountSettingsFields & getPasswordModifiedFields
These two methods are used at different stages of the process and for
different reasons. Separating them allows us to only execute them when
they are needed and more importantly _when_ they are needed.

eg. the getAccountSettingFiles is not required for bc users.

Remove the configuration for the `xs` prop, the only use of it is `12`.

Translate the passwordFields directly upon receiving them, as opposed to
doing it in the component.

Hard code the list of fields directly in the method, we had a method for
indirection that wasn't needed.

Bonus: flip all the `!isBcUser` ternaries

B2B-2158