refactor(account): B2B-2146 Use local instead of global state for login after registration
The login after registration functionality was trying to find email &
password on the global "registration" state. Since the state is quite
big and convoluted it meant looping through all the fields to find the
customer's email and password.

By explicitly exposing it we remove the dependency on global state and
make the code simpler.