fix(orders): B2BCAT-39 Honour orderBy when fetching results for OrderDetail pagination
We were not honouring the sort order when fetching the list of orders
for the top right corner pagination. This caused weird glitches if the
order list was sorted and we navigated to order details.

e.g. sorting by status and picking the last item on a page will cause the
current index not to match the real index on the list, making next/back
buttons to misbehave.