feat(register): B2B-3501 Honour `?redirectTo=checkout` after registration
This commit redirects the user back to checkout when the
`redirectTo=checkout` param is present.

This enhances the flow where a user click "create account" in checkout,
as it takes them back to where they were after registration.

As part of this same change, we've tighten the detection of certain
scenarios, making sure we avoid certain false~positives when detecting
checkout/change_password/etc.