feat(register): B2B-3501 Honour `?redirectTo=check-out after registration (second try)
This commit redirects the user back to checkout when the
`redirectTo=check-out` param is present.

This enhances the flow where a user click "create account" in checkout,
as it takes them back to where they were after registration.

As part of this same change, we've tighten the detection of certain
scenarios, making sure we avoid certain false~positives when detecting
checkout/change_password/etc.

This is the second attempt at trying to honour the `redirectTo` query
param in buyer portal to redirect users back to checkout.

In the previous attempt, the pathname was set to `/` even if there was
no original hash in the url, this caused a `/#/` to always show.

This commit minimised the amount of change to `routes/index.tsx` but
ensures the query param is usable.