feat(quotes): B2B-3328 use backend product validation is feature flag value is true (#554)
## What/Why?
- I'm using the product backend validations feature flag in the Quote
Draft to avoid the products quantity and non purchasable validations in
the front end to cover the scenario of backordered products. In case we
have errors returned by the backend we will render a snackbar error to
let know users the products that not available to be added to a quote.
- I'm adding a new quoteBuilder file to store the quote builders that
can be reused across different spec files.

## Rollout/Rollback
Revert

## Testing

Adding product to quote:


https://github.com/user-attachments/assets/71f9caca-db8f-4258-8033-591d9728a806

Rendering snackbar if api returns with product validation error:


https://github.com/user-attachments/assets/b704350d-8203-4688-8a4c-b5a6f4255636