feat(quotes): B2B-3554 getting available to sell and unlimited backorder values from backend (#564)
## What/Why?

- I'm using the new availableToSell and umlimitedBackorder keys from
searchProducts to update the soft validations strings in the QuoteTable
and shopping list errored panel. Validations cover the different
scenarios such as Min and Max quantity values for each product, if is
out of stock and display the value of stock in case the user is setting
a bigger quantity of the current stock value.
- I also updated the current warning stock message in the QuoteTable to
use the availableToSell key so is compatible with backordered products.

## Rollout/Rollback
Revert

## Testing


https://github.com/user-attachments/assets/5236b4e5-e204-465d-a1d7-5280dfc4114b