fix(orders): B2B-3691 adding test case for successful cart creation when backend validation is enabled and product variant (#574)
## What/Why?
Adding test case to cover scenario when we are creating a cart
successfully from a purchased product with variants. This was covered
for the front end validations scenario but it was missing when using the
new backend validations feature flag.

## Rollout/Rollback
Revert

## Testing
<img width="866" height="705" alt="image"
src="https://github.com/user-attachments/assets/c039572a-e444-411e-b8fe-cfda5f4c3367"
/>