fix(shopping-lists): B2B-3744 Cleanup messages before adding to cart
We keep a list of all the products that error/succeed when adding to the
cart.

This list is used to render the ReAddToCart modal and display error
messages in it.

Make sure the prior errors/successes are cleaned up before trying to
add new products to the cart, this prevents old "alerts" to linger
around.