fix(shopping-lists): B2B-3744 Cleanup messages before adding to cart (#603)
Jira: [B2B-3744](https://bigcommercecloud.atlassian.net/browse/B2B-3744)

⚠️ This PR is better looked at ignoring white spaces:
![https://github.com/bigcommerce/b2b-buyer-portal/pull/603/files?w=1](https://github.com/bigcommerce/b2b-buyer-portal/pull/603/files?w=1)

## What/Why?
We keep a list of all the products that error/succeed when adding to the
cart.

This list is used to render the ReAddToCart modal and display error
messages in it.

Make sure the prior errors/successes are cleaned up before trying to add
new products to the cart, this prevents old "alerts" to linger around.

## Rollout/Rollback
Revert

## Testing
### Before:


https://github.com/user-attachments/assets/cc7b1f72-e668-405d-8ba9-601a63b46693


### After:


https://github.com/user-attachments/assets/2a82e49b-ea58-4b00-b234-f3b335a99aa7



[B2B-3744]:
https://bigcommercecloud.atlassian.net/browse/B2B-3744?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ