fix(quotes): B2B-3925 load quotes details with backend validations (#616)
## What/Why?

When loading a quote with the backend validations feature flag enabled
we are still using front end validations in the QuoteDetail component to
know if there's any out of stock or non purchasable error in the
product, so we need to differentiate using the feature flag when is
required to use front end or backend validations.

For products with a warning status, we should be able to checkout the
quote and see no errors in the page.
For products with a error status, we should see a snackbar for each
failed product and proceed to checkout button should be disable.

## Rollout/Rollback
Revert

## Testing


https://github.com/user-attachments/assets/36f0da33-6f1a-466c-aed6-8d41b5e404c0


https://github.com/user-attachments/assets/7c024b5a-5c69-4fc0-9f04-915febb48000


https://github.com/user-attachments/assets/e64f216a-7352-42da-be72-ec5591c2d23d