fix(quotes): B2B-3964 showing TBD only for products with warning or erros (#639)
## What/Why?
The quote detail summary is currently showing TBD for all prices even
after submission of the quote since we are relying only in the front end
flow validations to pass the "showPrice" prop to the quote summary.

When having a purchasable product we should show TBD only on Shipping,
only after revision we should see the value of the shipping.

When having out of stock/non purchasable products, we should display TBD
on all prices and only render the prices after revision of the quote.

## Rollout/Rollback
Revert

## Testing


https://github.com/user-attachments/assets/493d2ba0-260c-498f-830f-d176e2a03c33