fix(orders): B2B-3942 lift reorder checkbox state to parent (#662)
Jira: [B2B-3942](https://bigcommercecloud.atlassian.net/browse/B2B-3942)

## What/Why?
Backend validation needs to uncheck any reorder items that were
successfully added to the cart. To do that reliably, the checkbox state
has to live in the parent (`OrderDialog`) instead of inside
`OrderCheckboxProduct`.

This PR just lifts that state so the parent can eventually handle the
success case and drive the checkboxes accordingly.

Note: This lays some of the groundwork for a later PR where we actually
do the unchecking on success
<!--
  A description about what this pull request implements and its purpose.
Try to be detailed and describe any technical details to simplify the
job
  of the reviewer and the individual on production support.
-->

## Rollout/Rollback
revert and redeploy
<!--
Detail how this change will be rolled out. Include reference to any
experiments and how the success will be measured as the experiment is
ramped.

Document rollback procedures. Is rolling back the change as simple as
rolling back an experiment or does it require reverting code? Are there
database migrations that may change our decision to roll forward instead
of
back?
-->

## Testing
Should be no change in behavior with checkbox functionality


https://github.com/user-attachments/assets/31323e23-2283-4e64-bbbe-944c738b6171


<!--
Provide as much information as you can about how you tested and how
another
Engineer can test your change. Include screenshots, or test run output
where appropriate.
-->


[B2B-3942]:
https://bigcommercecloud.atlassian.net/browse/B2B-3942?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ